refactor(llm): split verification helpers
